+++ gobgp/cmd/bmp.go
@@ -18,7 +18,7 @@ package cmd
 import (
 	"fmt"
 	api "github.com/osrg/gobgp/api"
-	"github.com/osrg/gobgp/packet"
+	"github.com/osrg/gobgp/packet/bgp"
 	"github.com/spf13/cobra"
 	"golang.org/x/net/context"
 	"net"
