+++ gobgp/cmd/common.go
@@ -20,7 +20,7 @@ import (
 	"encoding/json"
 	"fmt"
 	"github.com/osrg/gobgp/api"
-	"github.com/osrg/gobgp/packet"
+	"github.com/osrg/gobgp/packet/bgp"
 	"google.golang.org/grpc"
 	"net"
 	"os"
