+++ gobgp/cmd/global_test.go
@@ -17,7 +17,7 @@ package cmd
 
 import (
 	"fmt"
-	"github.com/osrg/gobgp/packet"
+	"github.com/osrg/gobgp/packet/bgp"
 	"github.com/stretchr/testify/assert"
 	"strings"
 	"testing"
