+++ openswitch/openswitch.go
@@ -20,7 +20,7 @@ import (
 	log "github.com/Sirupsen/logrus"
 	api "github.com/osrg/gobgp/api"
 	"github.com/osrg/gobgp/gobgp/cmd"
-	"github.com/osrg/gobgp/packet"
+	"github.com/osrg/gobgp/packet/bgp"
 	"github.com/osrg/gobgp/server"
 	ovsdb "github.com/osrg/libovsdb"
 	"github.com/satori/go.uuid"
