+++ packet/bgp/bgpattrtype_string.go
