+++ server/rpki_test.go
@@ -23,7 +23,7 @@ import (
 
 	"github.com/armon/go-radix"
 	"github.com/osrg/gobgp/config"
-	"github.com/osrg/gobgp/packet"
+	"github.com/osrg/gobgp/packet/bgp"
 	"github.com/stretchr/testify/assert"
 )
