+++ table/destination_test.go
@@ -17,7 +17,7 @@ package table
 
 import (
 	//"fmt"
-	"github.com/osrg/gobgp/packet"
+	"github.com/osrg/gobgp/packet/bgp"
 	"github.com/stretchr/testify/assert"
 	"net"
 	"testing"
