+++ table/message_test.go
@@ -16,7 +16,7 @@
 package table
 
 import (
-	"github.com/osrg/gobgp/packet"
+	"github.com/osrg/gobgp/packet/bgp"
 	"github.com/stretchr/testify/assert"
 	"testing"
 	"time"
