+++ table/table_manager_test.go
@@ -18,7 +18,7 @@ package table
 import (
 	_ "fmt"
 	log "github.com/Sirupsen/logrus"
-	"github.com/osrg/gobgp/packet"
+	"github.com/osrg/gobgp/packet/bgp"
 	"github.com/stretchr/testify/assert"
 	"net"
 	"os"
